feat(compiler): Make trip counts and memory usage optional in statistics passes
This makes the trip counts of operations in the TFHE statistics pass
as well as the per-location memory usage statistics in the memory
usage statistics pass optional. These values are unset if the trip
count could not be determined statically.

+++ compilers/concrete-compiler/compiler/lib/Support/CompilationFeedback.cpp
@@ -194,7 +194,7 @@ llvm::json::Value toJSON(const mlir::concretelang::CompilationFeedback &v) {
       keysJson.push_back(std::move(keyJson));
     }
     statisticJson.insert({"keys", std::move(keysJson)});
-    statisticJson.insert({"count", (int64_t)statistic.count});
+    statisticJson.insert({"count", statistic.count});
 
     statisticsJson.push_back(std::move(statisticJson));
   }
